+++ lib/roo_on_rails.rb
@@ -12,4 +12,5 @@ if defined?(Rails)
   require 'roo_on_rails/railties/sidekiq'
   require 'roo_on_rails/railties/rake_tasks'
   require 'roo_on_rails/railties/google_oauth'
+  require 'roo_on_rails/railties/routemaster'
 end